Update examples and tutorial
