+++ docs/experimental/viz_toon.py
@@ -2,6 +2,7 @@ import numpy as np
 from fury import window, actor, ui, io, utils
 from fury.data.fetcher import fetch_viz_models, read_viz_models
 import vtk
+from fury.shaders import shader_to_actor
 
 fetch_viz_models()
 dragon = read_viz_models('dragon.obj')
@@ -11,15 +12,8 @@ dragon = utils.get_actor_from_polymapper(dragon)
 
 dragon.GetProperty().SetDiffuse(0.7)
 
-mapper = dragon.GetMapper()
-
-mapper.AddShaderReplacement(
-    vtk.vtkShader.Fragment,
-    '//VTK::Light::Impl',
-    True,
+toon_shader = \
     """
-    //VTK::Light::Impl
-
     vec4 color;
 
     float ll = length(diffuse);
@@ -33,9 +27,9 @@ mapper.AddShaderReplacement(
         color = vec4(0.2,0.1,0.1,1.0);
 
     fragOutput0 = color;
-    """,
-    False
-)
+    """
+
+shader_to_actor(dragon, "fragment", block="light", impl_code=toon_shader)
 
 scene = window.Scene()
 scene.background((1.0, 0.8, 0.8))
